Socket basic
