+++ src/kr/ac/snu/sbkim28/Main.java
@@ -7,12 +7,12 @@ import kr.ac.snu.sbkim28.data.korean.KoreanWordReader;
 import kr.ac.snu.sbkim28.data.korean.KoreanWordSelector;
 import kr.ac.snu.sbkim28.game.core.ConsoleInputPlayer;
 import kr.ac.snu.sbkim28.game.core.Environment;
+import kr.ac.snu.sbkim28.game.core.LocalEnvironment;
 import kr.ac.snu.sbkim28.game.core.Player;
 
 import java.io.BufferedWriter;
 import java.io.IOException;
 import java.io.OutputStreamWriter;
-import java.util.*;
 
 public class Main {
     public static final BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
@@ -26,7 +26,7 @@ public class Main {
 
         Player player1 = new ConsoleInputPlayer();
         Player player2 = new ConsoleInputPlayer();
-        Environment environment = new Environment(player1, player2);
+        Environment environment = new LocalEnvironment(player1, player2);
         environment.setWordSet(kws.getWords());
         environment.runGame();
     }
