+++ src/kr/ac/snu/sbkim28/game/core/Player.java
@@ -5,20 +5,20 @@ package kr.ac.snu.sbkim28.game.core;
  * @author sbkim28
  * @version 1.0
  * @see ConsoleInputPlayer
- * @see Environment
+ * @see LocalEnvironment
  */
 public interface Player {
 
     /**
      * 플레이어에게 자신의 턴이 왔음을 알려줌.
-     * {@link Environment}에서 해당 플레이어의 턴이 올 때 호출됨.
+     * {@link LocalEnvironment}에서 해당 플레이어의 턴이 올 때 호출됨.
      * @param c 현재 글자.
      */
     void notifyTurn(char c);
 
     /**
      * 플레이어에게 자신의 턴이 왔음을 알려줌.
-     * {@link Environment}에서 해당 플레이어의 턴이 올 때 호출됨.
+     * {@link LocalEnvironment}에서 해당 플레이어의 턴이 올 때 호출됨.
      * @param c 현재 글자.
      * @param sub 두음법칙을 적용할 수 있는 경우의 글자.
      */
@@ -37,14 +37,14 @@ public interface Player {
 
     /**
      * 사용자가 입력한 단어가 성공하였는지 여부를 알려줌.
-     * {@link Environment}에서 {@link Player#getWord()}를 호출한 이후
+     * {@link LocalEnvironment}에서 {@link Player#getWord()}를 호출한 이후
      * 결과를 이 method를 통해서 전달함.
      */
     void notifySuccess(boolean success);
 
     /**
      * 게임이 종료되었는지 호출함.
-     * {@link Environment}에서 시간 초과 등의 이유로 게임이 끝났을 경우
+     * {@link LocalEnvironment}에서 시간 초과 등의 이유로 게임이 끝났을 경우
      * 이 메써드가 호출됨.
      * 이 메써드가 호출되면 {@link #getWord()} 대기를 중단하여야 하며,
      * {@link #getWord()}에서는 null을 반환하여야 함.
