+++ src/kr/ac/snu/sbkim28/game/core/SocketPlayer.java
@@ -1,0 +1,19 @@
+package kr.ac.snu.sbkim28.game.core;
+
+public class SocketPlayer extends AbstractPlayer {
+
+    @Override
+    public String getWord() {
+        return null;
+    }
+
+    @Override
+    public void notifySuccess(boolean success) {
+
+    }
+
+    @Override
+    public void notifyGameOver() {
+
+    }
+}
